Fixed [1526529] Title Url should be empty, not null.
Fixed some fubar'd "TitleUrl" handling.
Renamed the property TitleUrl to AlternativeTitleUrl.  Then added a new TitleUrl property which is a read only property.  TitleUrl is what the UI controls should bind to for displaying the URL in the title of a post only.  AlternativeTitleUrl is what is set in the admin tool to indicate that the blog entry title should link to something other than the blog post.

I changed a lot of code so I need to test this carefully.