Making script changes in preparation for merging the SQL Membership Provider.
Added Installation.01.99.90.sql - This is a bit of a hack in that we'll never have a real 01.99.90 version of subtext.  But by placing this file here it'll get run as part of the upgrade to 02.00.00.  02.00.00 will contain the select statements to copy data into the new membership tables.

StoredProcedures.sql contains the new membership stored procs.