Can now login to HostAdmin on a fresh install. Fixes #1
The problem here is when we logged in, we were setting the wrong cookie. We need to make sure the host admin cookie name is correctly set.
Also fixed it so that logging in as an admin, who also happens to be the host admin, logs into both.
Also fixed ability to logout of HostAdmin. Bam!