Support "placeholder" Gravatar image in blog skins
Only change the Visible property of the Gravatar image to true when an email address is specified in the comment. This allows a skin to specify a default Gravatar image (e.g. PlaceHolderImage="~/Skins/TechnologyToolbox1/Images/Silhouette-1.jpg") but also choose to hide the default image when no email address is available to retrieve a Gravatar (i.e. by specifying Visible="false" on the Image control).